# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API documentation for Zenick Aggregator.",
        "title": "Zenick Aggregator API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
        "/histories": {
            "get": {
                "description": "Retrieve detailed history records with related entities",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Histories"
                ],
                "summary": "Get detailed histories",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.HistoryResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/histories/GetHistoryByCondition": {
            "get": {
                "description": "Retrieve history based on filter conditions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Histories"
                ],
                "summary": "Get history by condition",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Provider name",
                        "name": "provider",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Token name",
                        "name": "token",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Operation name",
                        "name": "operation",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.HistoryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/liquidityPoolHistories": {
            "get": {
                "description": "Retrieve detailed liquidityPoolHistory records with related entities",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "LiquidityPoolHistories"
                ],
                "summary": "Get detailed liquidityPoolHistories",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.LiquidityPoolHistoryResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/liquidityPoolHistories/GetLiquidityPoolHistoryByCondition": {
            "get": {
                "description": "Retrieve Liquidity Pool History based on filter conditions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "LiquidityPoolHistories"
                ],
                "summary": "Get Liquidity Pool History by condition",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Provider name",
                        "name": "provider",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.LiquidityPoolHistoryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "dto.HistoryResponse": {
            "type": "object",
            "properties": {
                "apr": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "operation": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "dto.LiquidityPoolHistoryResponse": {
            "type": "object",
            "properties": {
                "apr": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "token_a": {
                    "type": "string"
                },
                "token_b": {
                    "type": "string"
                }
            }
        }
    }
}